+++ app/src/main/java/org/simple/clinic/facility/change/FacilityChangeActivity.kt
@@ -7,28 +7,25 @@ import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import io.github.inflationx.viewpump.ViewPumpContextWrapper
 import io.reactivex.Observable
-import io.reactivex.subjects.PublishSubject
+import io.reactivex.rxkotlin.ofType
 import kotlinx.android.synthetic.main.screen_facility_change.*
 import org.simple.clinic.ClinicApp
 import org.simple.clinic.R
-import org.simple.clinic.bindUiToController
+import org.simple.clinic.ReportAnalyticsEvents
 import org.simple.clinic.di.InjectorProviderContextWrapper
 import org.simple.clinic.facility.Facility
 import org.simple.clinic.facility.change.confirm.ConfirmFacilityChangeSheet
 import org.simple.clinic.facility.change.confirm.FacilityChangeComponent
+import org.simple.clinic.mobius.MobiusDelegate
 import org.simple.clinic.util.LocaleOverrideContextWrapper
 import org.simple.clinic.util.RuntimePermissions
+import org.simple.clinic.util.unsafeLazy
 import org.simple.clinic.util.wrap
-import org.simple.clinic.widgets.ScreenCreated
-import org.simple.clinic.widgets.ScreenDestroyed
 import org.simple.clinic.widgets.UiEvent
 import java.util.Locale
 import javax.inject.Inject
 
-class FacilityChangeActivity : AppCompatActivity() {
-
-  @Inject
-  lateinit var controller: FacilityChangeActivityController
+class FacilityChangeActivity : AppCompatActivity(), FacilityChangeUi, FacilityChangeUiActions {
 
   @Inject
   lateinit var locale: Locale
@@ -36,23 +33,35 @@ class FacilityChangeActivity : AppCompatActivity() {
   @Inject
   lateinit var runtimePermissions: RuntimePermissions
 
-  private val onDestroys = PublishSubject.create<ScreenDestroyed>()
+  @Inject
+  lateinit var effectHandlerFactory: FacilityChangeEffectHandler.Factory
+
+  private val events by unsafeLazy {
+    facilityClicks()
+        .compose(ReportAnalyticsEvents())
+        .share()
+  }
+
+  private val delegate by unsafeLazy {
+    val uiRenderer = FacilityChangeUiRenderer(this)
+
+    MobiusDelegate.forActivity(
+        events = events.ofType(),
+        defaultModel = FacilityChangeModel.create(),
+        update = FacilityChangeUpdate(),
+        effectHandler = effectHandlerFactory.create(this).build(),
+        init = FacilityChangeInit(),
+        modelUpdateListener = uiRenderer::render
+    )
+  }
 
   override fun onCreate(savedInstanceState: Bundle?) {
     super.onCreate(savedInstanceState)
     setContentView(R.layout.screen_facility_change)
 
-    bindUiToController(
-        ui = this,
-        events = Observable.merge(
-            screenCreates(),
-            facilityClicks()
-        ),
-        controller = controller,
-        screenDestroys = onDestroys
-    )
-
     setupUiComponents()
+
+    delegate.onRestoreInstanceState(savedInstanceState)
   }
 
   override fun attachBaseContext(baseContext: Context) {
@@ -66,9 +75,19 @@ class FacilityChangeActivity : AppCompatActivity() {
     super.attachBaseContext(wrappedContext)
   }
 
-  override fun onDestroy() {
-    onDestroys.onNext(ScreenDestroyed())
-    super.onDestroy()
+  override fun onStart() {
+    super.onStart()
+    delegate.start()
+  }
+
+  override fun onStop() {
+    delegate.stop()
+    super.onStop()
+  }
+
+  override fun onSaveInstanceState(outState: Bundle) {
+    delegate.onSaveInstanceState(outState)
+    super.onSaveInstanceState(outState)
   }
 
   private fun setupDi() {
@@ -83,8 +102,6 @@ class FacilityChangeActivity : AppCompatActivity() {
   private fun setupUiComponents() {
     facilityPickerView.backClicked = this@FacilityChangeActivity::finish
   }
-
-  private fun screenCreates() = Observable.just(ScreenCreated())
 
   private fun facilityClicks(): Observable<UiEvent> {
     return Observable.create { emitter ->
@@ -107,13 +124,13 @@ class FacilityChangeActivity : AppCompatActivity() {
     finish()
   }
 
-  fun goBack() {
+  override fun goBack() {
     val intent = Intent()
     setResult(Activity.RESULT_CANCELED, intent)
     finish()
   }
 
-  fun openConfirmationSheet(facility: Facility) {
+  override fun openConfirmationSheet(facility: Facility) {
     startActivityForResult(
         ConfirmFacilityChangeSheet.intent(this, facility),
         OPEN_CONFIRMATION_SHEET