+++ app/src/main/java/org/simple/clinic/facility/change/FacilityChangeEffect.kt
@@ -1,0 +1,11 @@
+package org.simple.clinic.facility.change
+
+import org.simple.clinic.facility.Facility
+
+sealed class FacilityChangeEffect
+
+object LoadCurrentFacility: FacilityChangeEffect()
+
+data class OpenConfirmFacilityChangeSheet(val facility: Facility): FacilityChangeEffect()
+
+object GoBack: FacilityChangeEffect()