+++ app/src/main/java/org/simple/clinic/facility/change/FacilityChangeModel.kt
@@ -1,0 +1,24 @@
+package org.simple.clinic.facility.change
+
+import android.os.Parcelable
+import kotlinx.android.parcel.Parcelize
+import org.simple.clinic.facility.Facility
+
+@Parcelize
+data class FacilityChangeModel(
+    val currentFacility: Facility?
+) : Parcelable {
+
+  companion object {
+    fun create(): FacilityChangeModel {
+      return FacilityChangeModel(currentFacility = null)
+    }
+  }
+
+  val hasLoadedCurrentFacility: Boolean
+    get() = currentFacility != null
+
+  fun currentFacilityLoaded(facility: Facility): FacilityChangeModel {
+    return copy(currentFacility = facility)
+  }
+}